Ability to provide html attributes to fieldset legends
This PR will add the ability to propagate html attributes down to the legend
inner element (`span` at the moment) for the `radio_button_fieldset` and
`check_box_fieldset` methods.

This solves an accessibility issue we were facing: legends need to exist for
screen reader users, but we wanted them to be visually hidden for non screen
reader users, and this was not possible with the current version of the gem.

Now, you can provide an optional hash as in these examples:

```ruby
f.radio_button_fieldset :case_type, choices: [...],
  legend_options: {class: 'visuallyhidden'}

f.check_box_fieldset :waste_transport, [...],
  legend_options: {class: 'visuallyhidden'}
```

And the legend element will produce the following markup:

```html
<legend>
  <span class="visuallyhidden">What is your appeal about?</span>
</legend>
```

Or, when errors found:

```html
<legend>
  <span class="visuallyhidden">What is your appeal about?</span>
  <span class="error-message">Select what your appeal is about</span>
</legend>
```

Note the `legend_options`, if provided, will overwrite the current span class
`form-label-bold`. If not provided, the current behaviour remains unchanged.

The most common scenario if `legend_options` is provided is to customize the
span so overwriting the hardcoded class is preferible. If you want to maintain
the `form-label-bold` AND provide extra attributes, make sure to add it to your
options, i.e. `legend_options: {class: 'form-label-bold color-red', id: 'test'}`

This also ensures backward-compatibility.