Merge attributes without overwriting defaults
Based on recent feedback on this quite old PR, it was decided to not
overwrite the default legend span class ('form-label-bold') when another
class is provided, for example to make it visually hidden.

With these amendments, any class or classes provided by
`legend_options` will be appended to the default one. It is still
possible to provide any other attributes, not only 'class' and these
will be propagated. An example in the tests:

```ruby
legend_options: {class: 'visuallyhidden', lang: 'en'}
```

will produce in the markup:

```html
class="form-label-bold visuallyhidden" lang="en"
```

Strings or arrays can be provided.

A bit of refactor to use this new method and simplify existing methods.