Make provision for #text_area being a special case
Like select boxes, text area is no longer simply a `.form-control`. It
is now classified as `.govuk-textarea`.

It may make sense to separate it completely, this will simplify much of
the code (and specs) considerably but is quite a big chunk of work as
the shared specs will need to be reworked into smaller shared groups.