Make the tests pass without repeating output
When using the form builder in a sample app, the output is duplicated
several times by `#safe_concat`. Using `#safe_join` makes it render
properly but in the tests the output is no longer captured properly.

There must be an actual underlying cause for this somewhere but for the
time being, wrap the `#check_box_input` calls in another safe_join.