Adds hint tags before error message
The design system specifies that error messages go after any hint tags.
https://design-system.service.gov.uk/components/error-message/
This commit fixes the form builder to match this requirement.

Rails' label method adds the error span to the label before we add in
the hints so we need to split on the span if it's present and add the
tags before that. If it's not present then we fall back to the old
behaviour.

It seems like the tag and element args to add_hint are always :label and
a label tag, but refactoring out these arguments is left as an exercise
to the reader.